+++ src/main/java/com/saditec/platform/error/ResponseExceptionHandler.java
@@ -1,0 +1,21 @@
+package com.saditec.platform.error;
+
+import com.saditec.platform.type.ApiResponse;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+@RestControllerAdvice
+public class ResponseExceptionHandler extends ResponseEntityExceptionHandler {
+
+    @ExceptionHandler({BadRequestException.class})
+    protected ResponseEntity<Object> handleBadRequestException(BadRequestException badRequestException) {
+        ApiResponse<Object> apiResponse = new ApiResponse<>();
+        apiResponse.setCode(badRequestException.hashCode());
+        apiResponse.setStatus(HttpStatus.BAD_REQUEST.value());
+        apiResponse.setMessage(badRequestException.getMessage());
+        return new ResponseEntity<>(apiResponse, HttpStatus.BAD_REQUEST);
+    }
+}