+++ src/main/java/com/saditec/platform/security/filter/JSONWebTokenAuthenticationFilter.java
@@ -1,0 +1,4 @@
+package com.saditec.platform.security.filter;
+
+public class JSONWebTokenAuthenticationFilter {
+}